% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhd_dir.R
\name{lhd_dir}
\alias{lhd_dir}
\title{Build project folder objects.}
\usage{
lhd_dir(project, subproject)
}
\arguments{
\item{project}{contains the top-level folder.}

\item{subproject}{contains the project-specific subfolder.}
}
\value{
The working directory as \code{working}.
}
\description{
Creates and setwd() to project directory top-level folder. Manage folders using file.path().
}
\examples{
library(devtools)
install_github("JECheadle/RLifeHD",
    auth_token = "b1ab72ab80f8caf23999bd91cf4042018efaea18")

library(RLifeHD)
Wdir <- lhd_dir("StudentHD1", "sHD1_Race")
Wdir <- lhd_dir(StudentHD1, sHD1_Race)
}
