% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhd_dir_builder.R
\name{lhd_dir_builder}
\alias{lhd_dir_builder}
\title{Check project, user, and directory data files, resave from master if needed.}
\usage{
lhd_dir_builder()
}
\arguments{
\item{Takes}{no parameters.}
}
\description{
\code{lhd_dir_builder} will check the master excel file against the *.RData
version. If they differ will save new versions of RData and *.csv versions
for the the stata version.
}
\examples{
## Run the command
lhd_dir_builder()

}
