% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhd_rowMean.R
\name{lhd_rowMean}
\alias{lhd_rowMean}
\title{Calculates row missing with means coded to NA based on number of missing variables.}
\usage{
lhd_rowMean(data, newvar, vars, num_miss)
}
\arguments{
\item{data}{the tibble or data.frame.}

\item{newvar}{is the name of the new variable.}

\item{vars}{is the vector of variable names that will be used to calculate the row mean.}

\item{num_miss}{is the number of missing that will be used to determine if the row mean should be missing. It is expressed as \code{>=}. Therefore, if \code{num_miss == 2}, the variable would be coded to \code{NA} if 2 or more variables in the list are missing.}
}
\value{
The tibble or data.frame with the row mean and the number of missing across the variable list.
}
\description{
Calculates row missing with means coded to NA based on number of missing variables.
}
\examples{
library(RLifeHD)
library(tidyverse)
master <- tibble(x = c(1:5, NA,NA), y = c(1,2,3,4,NA, NA,7))
master
vars <- c("x", "y")
master <- lhd_rowMean(master, my_row_mean, vars, 1)
master
}
